chore(agent,controlplane): error, logging, and mild code cleanup
