+++ crates/agent/src/client.rs
@@ -105,8 +105,18 @@ pub async fn ws_connection(ws_req: Request, state: Arc<GlobalState>) {
 
             // handle outgoing responses
             msg = server_response_out.recv() => {
-                let msg = msg.expect("internal RPC channel closed");
-                let bin = bincode::serialize(&control::MuxedMessageOutgoing::Child(msg)).expect("failed to serialize response");
+                let Some(msg) = msg else {
+                    error!("internal agent RPC channel closed");
+                    break;
+                };
+                let bin = match bincode::serialize(&control::MuxedMessageOutgoing::Child(msg)) {
+                    Ok(bin) => bin,
+                    Err(e) => {
+                        error!("failed to serialize response: {e}");
+                        continue;
+                    }
+                };
+
                 let send = stream.send(tungstenite::Message::Binary(bin));
                 if tokio::time::timeout(Duration::from_secs(10), send).await.is_err() {
                     error!("The connection to the control plane was interrupted while sending agent message");
@@ -116,8 +126,17 @@ pub async fn ws_connection(ws_req: Request, state: Arc<GlobalState>) {
 
             // handle outgoing requests
             msg = client_request_out.recv() => {
-                let msg = msg.expect("internal RPC channel closed");
-                let bin = bincode::serialize(&control::MuxedMessageOutgoing::Parent(msg)).expect("failed to serialize request");
+                let Some(msg) = msg else {
+                    error!("internal agent RPC channel closed");
+                    break;
+                };
+                let bin = match bincode::serialize(&control::MuxedMessageOutgoing::Parent(msg)) {
+                    Ok(bin) => bin,
+                    Err(e) => {
+                        error!("failed to serialize request: {e}");
+                        continue;
+                    }
+                };
                 let send = stream.send(tungstenite::Message::Binary(bin));
                 if tokio::time::timeout(Duration::from_secs(10), send).await.is_err() {
                     error!("The connection to the control plane was interrupted while sending control message");
@@ -174,8 +193,18 @@ pub async fn ws_connection(ws_req: Request, state: Arc<GlobalState>) {
                     };
 
                     match msg {
-                        control::MuxedMessageIncoming::Child(msg) => server_request_in.send(msg).expect("internal RPC channel closed"),
-                        control::MuxedMessageIncoming::Parent(msg) => client_response_in.send(msg).expect("internal RPC channel closed"),
+                        control::MuxedMessageIncoming::Child(msg) => {
+                            if let Err(e) = server_request_in.send(msg) {
+                                error!("internal agent RPC channel closed: {e}");
+                                break;
+                            }
+                        },
+                        control::MuxedMessageIncoming::Parent(msg) => {
+                            if let Err(e) = client_response_in.send(msg) {
+                                error!("internal agent RPC channel closed: {e}");
+                                break;
+                            }
+                        }
                     }
                 }
